DMV-40 bug fixed
SVN: 15403

+++ common/source/java/ch/systemsx/cisd/common/filesystem/FileUtilities.java
@@ -713,7 +713,7 @@ public final class FileUtilities
         {
             return true;
         }
-        if (Unix.isOperational())
+        if (file.exists() && Unix.isOperational())
         {
             Unix.setAccessMode(file.getPath(), (short) 0777);
             return file.delete();
